+++ src/SDL/Event.hs
@@ -410,7 +410,9 @@ data ControllerButtonEventData =
 
 -- | Controller device event information
 data ControllerDeviceEventData =
-  ControllerDeviceEventData {controllerDeviceEventWhich :: !Int32
+  ControllerDeviceEventData {controllerDeviceEventConnection :: !ControllerDeviceEventConnection
+                             -- ^ Was the device added, removed, or remapped?
+                            ,controllerDeviceEventWhich :: !Int32
                              -- ^ The joystick instance ID that reported the event.
                             }
   deriving (Eq,Ord,Generic,Show,Typeable)
@@ -661,8 +663,8 @@ convertRaw (Raw.ControllerButtonEvent _ ts a b c) =
              (ControllerButtonEventData a 
                                         (fromNumber $ fromIntegral b)
                                         (fromNumber c))))
-convertRaw (Raw.ControllerDeviceEvent _ ts a) =
-  return (Event ts (ControllerDeviceEvent (ControllerDeviceEventData a)))
+convertRaw (Raw.ControllerDeviceEvent t ts a) =
+  return (Event ts (ControllerDeviceEvent (ControllerDeviceEventData (fromNumber t) a)))
 convertRaw (Raw.AudioDeviceEvent Raw.SDL_AUDIODEVICEADDED ts a b) =
   return (Event ts (AudioDeviceEvent (AudioDeviceEventData True a (b /= 0))))
 convertRaw (Raw.AudioDeviceEvent Raw.SDL_AUDIODEVICEREMOVED ts a b) =